WIP: try out shopify_gem
The generator did not create all the resources :/ Once @fomentia
discovered the proper way to create the state token (go through `/login`
so that omniauth could create and register it), and we had all the rest
of the resources for the app, authentication began to work.

Now we are having a Chrome problem with https pages loading http iframe
sources. It converts the http://localhost to https://localhost so that
our server complains about an invalid protocol.